Module spinner::menu

Module defining the menu part of this crate

Structs

Menu
A menu that has not yet been displayed

MenuOption
An individual MenuOption, check out the crate documentation on how to use it

Enums

MenuOptional
An enum specifying if a given field is optional or not

MenuType
An enum specifying the type of information the user should put in.

MenuValue
The value of the given menu, when given as an argument to the constructor this will be used as the default value. When returned after displaying the menu it will contain the information input by the user.