Struct spinner::menu::MenuOption

pub struct MenuOption(pub String, pub MenuType, pub Option<MenuValue>);

An individual MenuOption, check out the crate documentation on how to use it

Methods

impl MenuOption

fn get_string(self) -> Option<String>

Get the string if there is one, will panic if MenuType is not a string

Panic

This will panic if the types do not match

fn get_int(self) -> Option<i64>

Get the integer if there is one, will panic if MenuType is not an integer

Panic

This will panic if the types do not match

fn get_float(self) -> Option<f64>

Get the float if there is one, will panic if MenuType is not a float

Panic

This will panic if the types do not match