Struct spinner::SpinnerHandle

pub struct SpinnerHandle {
    // some fields omitted
}

A handle to the Spinner Thread.

Important, be sure to call close before dropping this struct to make sure the thread joins before the main thread might close. Otherwise you will get cutoff output.

Methods

impl SpinnerHandle

fn update(&self, st: String) -> Option<String>

Update the message that is given to the user as part of the spinner

Returns the String that is put in in case the sender could not send.

fn message(&self, msg: String) -> Option<String>

Print out a message above the Spinner for the user.

fn close(self)

Shutdown the thread and wait until it has joined.